backend/requirements.txt
requests
google-generativeai
IPython
flask
flask-cors
